+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- URL-escape trino cluster credentials ([#40], [#41]).
+
+[#40]: https://github.com/stackabletech/trino-lb/pull/40
+[#41]: https://github.com/stackabletech/trino-lb/pull/41
+
+
 ## [0.2.2] - 2024-06-28
 
 ### Fixed